[client] Regenerate protobuf bindings after rebase
